+++ grouping_flutter/lib/app/presentation/providers/theme_provider.dart
@@ -1,12 +1,12 @@
 import 'package:flutter/material.dart';
-import 'package:grouping_project/data/entities/theme_entity.dart';
+import 'package:grouping_project/core/theme_entity.dart';
 
 // TODO: 這是 ViewModel 不能存在在這裡
 
 class ThemeManager with ChangeNotifier {
   // 這是 ViewModel
   // brightness 用來切換主題是Model
-  ThemeModel theme = ThemeModel();
+  AppTheme theme = AppTheme();
   Brightness get brightness => theme.brightness;
   IconData get icon => theme.themeIcon;
   Widget get coverLogo => theme.coverLogo;
