+++ grouping_flutter/lib/core/message_model.dart
