+++ grouping_flutter/lib/dashboard/data/models/event_model.dart
@@ -1,6 +1,6 @@
 // ignore_for_file: unnecessary_this
-import 'package:grouping_project/data/repo/account_model.dart';
-import 'package:grouping_project/data/workspace/editable_card_model.dart';
+import 'package:grouping_project/dashboard/data/models/account_model.dart';
+import 'package:grouping_project/dashboard/data/models/editable_card_model.dart';
 
 // import 'account_model.dart';
